Remove deserialize
